pyproject.toml
[project]
name = "real-estate-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-adk>=0.4.0",
    "psycopg2-binary>=2.9.10",
    "sqlalchemy>=2.0.40",
    "sqlglot>=26.17.1",
]
